+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+* Updated
+  * Corrected to get width from bounding client rect instead of resize observer to avoid infinite loop.
+
+## 6.94.0 - (March 14, 2024)
+
+* Changed
+  * Minor dependency version bump.
+
 ## 6.93.0 - (March 8, 2024)
 
 * Changed